separated soylattedb library
